add accept: test user_login/logout/sign_up, auth user create/delete own quest and answers

+++ app/controllers/questions_controller.rb
@@ -9,8 +9,8 @@ class QuestionsController < ApplicationController
   end
 
   def create
-    @question = Question.new(question_params)
-    
+    @question = current_user.questions.new(question_params)
+
     if @question.save
       redirect_to @question, notice: 'Your question successfully created.'
     else
@@ -27,8 +27,12 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author?(question)
+      question.destroy
+      redirect_to questions_path, notice: "Question successfully delete"
+    else
+      redirect_to question, notice: 'You are not the author question.'
+    end
   end
 
   private
